# Dynamically push table filters from Top-N operator