--- a/src/execution/physical_plan/plan_top_n.cpp
+++ b/src/execution/physical_plan/plan_top_n.cpp
@@ -9,8 +9,9 @@ unique_ptr<PhysicalOperator> PhysicalPlanGenerator::CreatePlan(LogicalTopN &op) 
 
 	auto plan = CreatePlan(*op.children[0]);
 
-	auto top_n = make_uniq<PhysicalTopN>(op.types, std::move(op.orders), NumericCast<idx_t>(op.limit),
-	                                     NumericCast<idx_t>(op.offset), op.estimated_cardinality);
+	auto top_n =
+	    make_uniq<PhysicalTopN>(op.types, std::move(op.orders), NumericCast<idx_t>(op.limit),
+	                            NumericCast<idx_t>(op.offset), std::move(op.dynamic_filter), op.estimated_cardinality);
 	top_n->children.push_back(std::move(plan));
 	return std::move(top_n);
 }