--- a/src/include/duckdb/execution/operator/join/join_filter_pushdown.hpp
+++ b/src/include/duckdb/execution/operator/join/join_filter_pushdown.hpp
@@ -46,7 +46,9 @@ struct JoinFilterPushdownFilter {
 };
 
 struct PushdownFilterTarget {
-	PushdownFilterTarget(LogicalGet &get, vector<JoinFilterPushdownColumn> columns_p) : get(get), columns(std::move(columns_p)) {}
+	PushdownFilterTarget(LogicalGet &get, vector<JoinFilterPushdownColumn> columns_p)
+	    : get(get), columns(std::move(columns_p)) {
+	}
 
 	LogicalGet &get;
 	vector<JoinFilterPushdownColumn> columns;