--- a/src/include/duckdb/execution/operator/order/physical_top_n.hpp
+++ b/src/include/duckdb/execution/operator/order/physical_top_n.hpp
@@ -12,6 +12,7 @@
 #include "duckdb/planner/bound_query_node.hpp"
 
 namespace duckdb {
+struct DynamicFilterData;
 
 //! Represents a physical ordering of the data. Note that this will not change
 //! the data but only add a selection vector.
@@ -21,11 +22,14 @@ public:
 
 public:
 	PhysicalTopN(vector<LogicalType> types, vector<BoundOrderByNode> orders, idx_t limit, idx_t offset,
-	             idx_t estimated_cardinality);
+	             shared_ptr<DynamicFilterData> dynamic_filter, idx_t estimated_cardinality);
+	~PhysicalTopN() override;
 
 	vector<BoundOrderByNode> orders;
 	idx_t limit;
 	idx_t offset;
+	//! Dynamic table filter (if any)
+	shared_ptr<DynamicFilterData> dynamic_filter;
 
 public:
 	// Source interface