--- a/src/include/duckdb/optimizer/join_filter_pushdown_optimizer.hpp
+++ b/src/include/duckdb/optimizer/join_filter_pushdown_optimizer.hpp
@@ -20,7 +20,10 @@ class JoinFilterPushdownOptimizer : public LogicalOperatorVisitor {
 public:
 	explicit JoinFilterPushdownOptimizer(Optimizer &optimizer);
 
+public:
 	void VisitOperator(LogicalOperator &op) override;
+	static void GetPushdownFilterTargets(LogicalOperator &op, vector<JoinFilterPushdownColumn> columns,
+					  vector<PushdownFilterTarget> &targets);
 
 private:
 	void GenerateJoinFilters(LogicalComparisonJoin &join);