--- a/src/include/duckdb/optimizer/join_filter_pushdown_optimizer.hpp
+++ b/src/include/duckdb/optimizer/join_filter_pushdown_optimizer.hpp
@@ -13,6 +13,7 @@
 
 namespace duckdb {
 class Optimizer;
+struct JoinFilterPushdownColumn;
 
 //! The JoinFilterPushdownOptimizer links comparison joins to data sources to enable dynamic execution-time filter
 //! pushdown