--- a/src/include/duckdb/optimizer/join_filter_pushdown_optimizer.hpp
+++ b/src/include/duckdb/optimizer/join_filter_pushdown_optimizer.hpp
@@ -14,6 +14,7 @@
 namespace duckdb {
 class Optimizer;
 struct JoinFilterPushdownColumn;
+struct PushdownFilterTarget;
 
 //! The JoinFilterPushdownOptimizer links comparison joins to data sources to enable dynamic execution-time filter
 //! pushdown