--- a/src/include/duckdb/optimizer/topn_optimizer.hpp
+++ b/src/include/duckdb/optimizer/topn_optimizer.hpp
@@ -20,6 +20,9 @@ public:
 	unique_ptr<LogicalOperator> Optimize(unique_ptr<LogicalOperator> op);
 	//! Whether we can perform the optimization on this operator
 	static bool CanOptimize(LogicalOperator &op);
+
+private:
+	void PushdownDynamicFilters(LogicalTopN &op);
 };
 
 } // namespace duckdb