--- a/src/include/duckdb/optimizer/topn_optimizer.hpp
+++ b/src/include/duckdb/optimizer/topn_optimizer.hpp
@@ -12,6 +12,7 @@
 
 namespace duckdb {
 class LogicalOperator;
+class LogicalTopN;
 class Optimizer;
 
 class TopN {