--- a/src/include/duckdb/planner/filter/dynamic_filter.hpp
+++ b/src/include/duckdb/planner/filter/dynamic_filter.hpp
@@ -21,6 +21,7 @@ struct DynamicFilterData {
 	bool initialized = false;
 
 	void SetValue(Value val);
+	void Reset();
 };
 
 class DynamicFilter : public TableFilter {