--- a/src/include/duckdb/planner/operator/logical_top_n.hpp
+++ b/src/include/duckdb/planner/operator/logical_top_n.hpp
@@ -12,6 +12,7 @@
 #include "duckdb/planner/logical_operator.hpp"
 
 namespace duckdb {
+struct DynamicFilterData;
 
 //! LogicalTopN represents a comibination of ORDER BY and LIMIT clause, using Min/Max Heap
 class LogicalTopN : public LogicalOperator {
@@ -19,15 +20,16 @@ public:
 	static constexpr const LogicalOperatorType TYPE = LogicalOperatorType::LOGICAL_TOP_N;
 
 public:
-	LogicalTopN(vector<BoundOrderByNode> orders, idx_t limit, idx_t offset)
-	    : LogicalOperator(LogicalOperatorType::LOGICAL_TOP_N), orders(std::move(orders)), limit(limit), offset(offset) {
-	}
+	LogicalTopN(vector<BoundOrderByNode> orders, idx_t limit, idx_t offset);
+	~LogicalTopN() override;
 
 	vector<BoundOrderByNode> orders;
 	//! The maximum amount of elements to emit
 	idx_t limit;
 	//! The offset from the start to begin emitting elements
 	idx_t offset;
+	//! Dynamic table filter (if any)
+	shared_ptr<DynamicFilterData> dynamic_filter;
 
 public:
 	vector<ColumnBinding> GetColumnBindings() override {