--- a/src/include/duckdb/planner/table_filter.hpp
+++ b/src/include/duckdb/planner/table_filter.hpp
@@ -31,7 +31,8 @@ enum class TableFilterType : uint8_t {
 	CONJUNCTION_AND = 4,     // AND of different filters
 	STRUCT_EXTRACT = 5,      // filter applies to child-column of struct
 	OPTIONAL_FILTER = 6,     // executing filter is not required for query correctness
-	IN_FILTER = 7            // col IN (C1, C2, C3, ...)
+	IN_FILTER = 7,           // col IN (C1, C2, C3, ...)
+	DYNAMIC_FILTER = 8       // dynamic filters can be updated at run-time
 };
 
 //! TableFilter represents a filter pushed down into the table scan.