--- a/src/optimizer/join_filter_pushdown_optimizer.cpp
+++ b/src/optimizer/join_filter_pushdown_optimizer.cpp
@@ -30,8 +30,8 @@ bool PushdownJoinFilterExpression(Expression &expr, JoinFilterPushdownColumn &fi
 	return true;
 }
 
-void GenerateJoinFiltersRecursive(LogicalOperator &op, vector<JoinFilterPushdownColumn> columns,
-                                  JoinFilterPushdownInfo &pushdown_info) {
+void JoinFilterPushdownOptimizer::GetPushdownFilterTargets(LogicalOperator &op, vector<JoinFilterPushdownColumn> columns,
+                                  vector<PushdownFilterTarget> &targets) {
 	auto &probe_child = op;
 	switch (probe_child.type) {
 	case LogicalOperatorType::LOGICAL_LIMIT:
@@ -43,7 +43,7 @@ void GenerateJoinFiltersRecursive(LogicalOperator &op, vector<JoinFilterPushdown
 	case LogicalOperatorType::LOGICAL_CROSS_PRODUCT:
 		// does not affect probe side - recurse into left child
 		// FIXME: we can probably recurse into more operators here (e.g. window, unnest)
-		GenerateJoinFiltersRecursive(*probe_child.children[0], std::move(columns), pushdown_info);
+		GetPushdownFilterTargets(*probe_child.children[0], std::move(columns), targets);
 		break;
 	case LogicalOperatorType::LOGICAL_UNNEST: {
 		auto &unnest = probe_child.Cast<LogicalUnnest>();
@@ -54,7 +54,7 @@ void GenerateJoinFiltersRecursive(LogicalOperator &op, vector<JoinFilterPushdown
 				return;
 			}
 		}
-		GenerateJoinFiltersRecursive(*probe_child.children[0], std::move(columns), pushdown_info);
+		GetPushdownFilterTargets(*probe_child.children[0], std::move(columns), targets);
 		break;
 	}
 	case LogicalOperatorType::LOGICAL_EXCEPT:
@@ -80,7 +80,7 @@ void GenerateJoinFiltersRecursive(LogicalOperator &op, vector<JoinFilterPushdown
 				child_columns.push_back(new_col);
 			}
 			// then recurse into the child
-			GenerateJoinFiltersRecursive(*child, std::move(child_columns), pushdown_info);
+			GetPushdownFilterTargets(*child, std::move(child_columns), targets);
 
 			// for EXCEPT we can only recurse into the first (left) child
 			if (probe_child.type == LogicalOperatorType::LOGICAL_EXCEPT) {
@@ -102,15 +102,7 @@ void GenerateJoinFiltersRecursive(LogicalOperator &op, vector<JoinFilterPushdown
 				return;
 			}
 		}
-		// pushdown info can be applied to this LogicalGet - push the dynamic table filter set
-		if (!get.dynamic_filters) {
-			get.dynamic_filters = make_shared_ptr<DynamicTableFilterSet>();
-		}
-
-		JoinFilterPushdownFilter get_filter;
-		get_filter.dynamic_filters = get.dynamic_filters;
-		get_filter.columns = std::move(columns);
-		pushdown_info.probe_info.push_back(std::move(get_filter));
+		targets.emplace_back(get, std::move(columns));
 		break;
 	}
 	case LogicalOperatorType::LOGICAL_PROJECTION: {
@@ -127,7 +119,7 @@ void GenerateJoinFiltersRecursive(LogicalOperator &op, vector<JoinFilterPushdown
 				return;
 			}
 		}
-		GenerateJoinFiltersRecursive(*probe_child.children[0], std::move(columns), pushdown_info);
+		GetPushdownFilterTargets(*probe_child.children[0], std::move(columns), targets);
 		break;
 	}
 	case LogicalOperatorType::LOGICAL_AGGREGATE_AND_GROUP_BY: {
@@ -144,7 +136,7 @@ void GenerateJoinFiltersRecursive(LogicalOperator &op, vector<JoinFilterPushdown
 				return;
 			}
 		}
-		GenerateJoinFiltersRecursive(*probe_child.children[0], std::move(columns), pushdown_info);
+		GetPushdownFilterTargets(*probe_child.children[0], std::move(columns), targets);
 		break;
 	}
 	default:
@@ -205,7 +197,20 @@ void JoinFilterPushdownOptimizer::GenerateJoinFilters(LogicalComparisonJoin &joi
 		return;
 	}
 	// recurse the query tree to find the LogicalGets in which we can push the filter info
-	GenerateJoinFiltersRecursive(*join.children[0], pushdown_columns, *pushdown_info);
+	vector<PushdownFilterTarget> pushdown_filter_targets;
+	GetPushdownFilterTargets(*join.children[0], pushdown_columns, pushdown_filter_targets);
+	for(auto &target : pushdown_filter_targets) {
+		auto &get = target.get;
+		// pushdown info can be applied to this LogicalGet - push the dynamic table filter set
+		if (!get.dynamic_filters) {
+			get.dynamic_filters = make_shared_ptr<DynamicTableFilterSet>();
+		}
+
+		JoinFilterPushdownFilter get_filter;
+		get_filter.dynamic_filters = get.dynamic_filters;
+		get_filter.columns = std::move(target.columns);
+		pushdown_info->probe_info.push_back(std::move(get_filter));
+	}
 
 	// Even if we cannot find any table sources in which we can push down filters,
 	// we still initialize the aggregate states so that we have the possibility of doing a perfect hash join