… src/include/duckdb/storage/serialization/table_filter.json
… src/planner/filter/CMakeLists.txt
… src/planner/filter/null_filter.cpp
… src/planner/operator/logical_top_n.cpp
… src/storage/serialization/serialize_table_filter.cpp
… src/storage/table/row_group.cpp